requirements.txt
annotated-types==0.6.0
asgiref==3.7.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2.6
django-environ==0.11.2
django-import-export==3.3.7
django-phonenumber-field==7.2.0
django-unfold==0.20.0
djangorestframework==3.14.0
dotty-dict==1.3.1
et-xmlfile==1.1.0
gitdb==4.0.11
GitPython==3.1.41
google-api-python-client==1.4.1
gunicorn==21.2.0
httplib2==0.22.0
idna==3.6
importlib-resources==6.1.1
Jinja2==3.1.3
markdown-it-py==3.0.0
MarkupPy==1.14
MarkupSafe==2.1.3
mdurl==0.1.2
oauth2client==4.1.3
odfpy==1.4.1
openpyxl==3.1.2
packaging==24.0
phonenumbers==8.13.22
psycopg2-binary==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.5.3
pydantic_core==2.14.6
Pygments==2.17.2
pyparsing==3.1.1
python-gitlab==4.4.0
python-semantic-release==8.7.0
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.7.0
rsa==4.9
sgmllib3k==1.0.0
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sqlparse==0.4.4
tablib==3.5.0
tomlkit==0.12.3
typing_extensions==4.8.0
uritemplate==4.1.1
urllib3==2.1.0
xlrd==2.0.1
xlwt==1.3.0